return 웹소켓으로 변경

+++ danusys-drone/src/main/java/com/danusys/web/drone/dto/response/Gps.java
@@ -14,7 +14,7 @@ public class Gps {
     private Double currentHeight; // 높이
     private int wpDist; // 목표까지의거리
     private String missionType; //미션 종류 (귀환,이륙,이동 등등)
-    private Double heading; //방향 0이 north
+    private int heading; //방향 0이 north
     private float airSpeed;
   //  private float airspeed; //수직속도
  //   private float groundspeed;//수평속도
